refactor: split metrics.py into returns/ratios/drawdown/stats (roadmap 5.1)
features/metrics.py (1782 lines) is split by concern:
- returns.py   : annual_return, perf_index/returns/strat, returns_strat, roll_annual_return
- ratios.py    : annual_volatility, sharpe, sortino, calmar, diversified_ratio + roll_*
- drawdown.py  : drawdown, mdd, mad + roll_*
- stats.py     : accuracy, directional_accuracy, z_score, roll_z_score
- _metrics_helpers.py : all private helpers + the _handler dicts

metrics.py is kept as a thin re-export aggregator (not deleted): deleting
it would break 6 importing modules and the Sphinx features.metrics.rst page
(currentmodule + xrefs). The public API (fynance.*, fynance.features.metrics.*),
every import path and the docs are therefore unchanged. perf_strat/returns_strat
(public but never in __all__) are re-exported explicitly for callers/tests.

Verified: ruff, pytest 294 + doctests (now running on each new module),
property/parity tests, sphinx-build -W (0 warnings).
