+++ pyproject.toml
@@ -96,6 +96,11 @@ ignore = ["E501"]
 # Dual Cython/Python files use star imports to pull Cython symbols
 "fynance/algorithms/browsers.py" = ["F403", "F405"]
 "fynance/features/metrics.py" = ["F403", "F405"]
+"fynance/features/returns.py" = ["F403", "F405"]
+"fynance/features/ratios.py" = ["F403", "F405"]
+"fynance/features/drawdown.py" = ["F403", "F405"]
+"fynance/features/stats.py" = ["F403", "F405"]
+"fynance/features/_metrics_helpers.py" = ["F403", "F405"]
 "fynance/features/momentums.py" = ["F403", "F405"]
 "fynance/features/roll_functions.py" = ["F403", "F405"]
 "fynance/features/scale.py" = ["F403", "F405"]
